se reestructuran las carpetas, se corrige la nomenclatura y se crea una interface DAO con tipo generico para poder implementarla de manera global para cada una de las clases correspondientes a las entidades de la base de datos
