--- src/main/java/com/cleaningservices/dao/ICategoriaProductoDAO.java
@@ -1,4 +1,0 @@
-package com.cleaningservices.dao;
-
-public interface ICategoriaProductoDAO {
-}
