+++ src/main/java/com/cleaningservices/dao/IDAO.java
@@ -1,0 +1,8 @@
+package com.cleaningservices.dao;
+
+public interface IDAO<T> {
+
+    public void insertar(T entidad);
+    public void mostrarData();
+
+}
