--- src/main/java/com/cleaningservices/dao/IMetodoPagoDAO.java
@@ -1,4 +1,0 @@
-package com.cleaningservices.dao;
-
-public interface IMetodoPagoDAO {
-}
