--- src/main/java/com/cleaningservices/dao/IProductoDAO.java
@@ -1,4 +1,0 @@
-package com.cleaningservices.dao;
-
-public interface IProductoDAO {
-}
