+++ src/main/java/com/cleaningservices/daoImpl/UsuarioDAOImpl.java
@@ -1,4 +1,4 @@
 package com.cleaningservices.daoImpl;
 
-public class OrdenesDAOImpl {
+public class UsuarioDAOImpl {
 }
