+++ src/main/java/com/cleaningservices/model/CategoriaProducto.java
@@ -1,0 +1,4 @@
+package com.cleaningservices.model;
+
+public class CategoriaProducto {
+}
