+++ src/main/java/com/cleaningservices/model/Empleados.java
@@ -1,0 +1,4 @@
+package com.cleaningservices.model;
+
+public class Empleados {
+}
