+++ src/main/java/com/cleaningservices/model/MetodoPago.java
@@ -1,0 +1,4 @@
+package com.cleaningservices.model;
+
+public class MetodoPago {
+}
