+++ src/main/java/com/cleaningservices/model/Producto.java
@@ -1,0 +1,4 @@
+package com.cleaningservices.model;
+
+public class Producto {
+}
