--- src/main/java/com/cleaningservices/models/Ordenes.java
@@ -1,4 +1,0 @@
-package com.cleaningservices.models;
-
-public class Ordenes {
-}
